Attempted Integration of Group 01's Weather Station
- Adapted components and tests from Group 01's code to produce
WeatherStation1, which interacts with the WeatherStationIntegrator

- Commented out DataRelay's console output when encountering a
FileNotFoundException

- Attempted to fix a potential synchronization issue with the
WeatherStationIntegrator where the integrator's setState method could
modify the weatherDataLists field around the same time that the Wireless
Console could extract the values from the weatherDataList

Important Note:
The method I used to extract the data from Group 01's sensors relies on
changing System.out's PrintStream to prevent Group 01's code from
printing to the console. This seems to cause anomalies with the console
output of the WirelessConsole.