Added tests for the Compass GUI
Most of the tests simply check if any exceptions occur when calling the
relevant methods of the Compass.

Note:
I had issues getting the tests to cover the paint(Graphics) method of
the Compass. The current method I have used is to essentially place the
Compass in a JFrame and do so without using EventQueue's invokeLater
method. It is unclear what issues this may cause for the other tests.